authdirにtestを移行した
